Universe and lattice power calculation now working properly in Python, but multithreading is now broken

+++ src/Cell.h
@@ -44,38 +44,38 @@ class Cell {
 protected:
 
     /** A static counter for the number of cells in a simulation */
-    static short int _n;
+    static int _n;
 
     /** A static counter for the number of cell clones in a simulation */
     static int _clone_n;
 
     /** A monotonically increasing unique ID for each cell created */
-    short int _uid;
+    int _uid;
 
     /** A user-defined ID for each cell created */
-    short int _id;
+    int _id;
 
     /** The type of cell (ie MATERIAL or FILL) */
     cellType _cell_type;
 
     /** The ID for the universe within which this cell resides */
-    short int _universe;
+    int _universe;
 
     /** Map of bounding surface ids (keys) to surface pointers (values). 
      *  Keys are +/- depending on the side of surface in which this cell
      *  resides. */
-    std::map<short int, Surface*> _surfaces;
+    std::map<int, Surface*> _surfaces;
 
 public:
     Cell();
-    Cell(short int universe, short int id=0);
+    Cell(int universe, int id=0);
     virtual ~Cell();
-    short int getUid() const;
-    short int getId() const;
+    int getUid() const;
+    int getId() const;
     cellType getType() const;
-    short int getUniverse() const;
-    short int getNumSurfaces() const;
-    std::map<short int, Surface*> getSurfaces() const;
+    int getUniverse() const;
+    int getNumSurfaces() const;
+    std::map<int, Surface*> getSurfaces() const;
 
     /**
      * @brief Return the number of flat source regions in this cell. 
@@ -85,8 +85,8 @@ public:
      */
     virtual int getNumFSRs() =0;
 
-    void setUniverse(short int universe);
-    void addSurface(short int halfspace, Surface* surface);
+    void setUniverse(int universe);
+    void addSurface(int halfspace, Surface* surface);
     void setSurfacePointer(Surface* surface);
 
     bool cellContainsPoint(Point* point);
@@ -116,13 +116,13 @@ class CellBasic: public Cell {
 private: 
 
     /** A pointer to the material filling this cell */
-    short int _material;
+    int _material;
 
     /** The number of rings sub-dividing this cell */
-    short int _num_rings;
+    int _num_rings;
 
     /** The number of sectors sub-dividing this cell */
-    short int _num_sectors;
+    int _num_sectors;
 
     std::vector<CellBasic*> _rings;
     std::vector<CellBasic*> _sectors;
@@ -132,16 +132,16 @@ private:
     void sectorize();
 
 public:
-    CellBasic(short int universe, short int material,
-	      int rings=0, int sectors=0, short int id=0);
+    CellBasic(int universe, int material,
+	      int rings=0, int sectors=0, int id=0);
 
-    short int getMaterial() const;
-    short int getNumRings();
-    short int getNumSectors();
+    int getMaterial() const;
+    int getNumRings();
+    int getNumSectors();
     int getNumFSRs();
 
-    void setNumRings(short int num_rings);
-    void setNumSectors(short int num_sectors);
+    void setNumRings(int num_rings);
+    void setNumSectors(int num_sectors);
     CellBasic* clone();
     std::vector<CellBasic*> subdivideCell();
 
@@ -158,16 +158,16 @@ class CellFill: public Cell {
 
 private:
     /** A pair of the ID and universe pointer filling this cell */
-    std::pair<short int, Universe*> _universe_fill;
+    std::pair<int, Universe*> _universe_fill;
 
 public:
-    CellFill(short int universe, short int universe_fill, short int id=0);
+    CellFill(int universe, int universe_fill, int id=0);
 
-    short int getUniverseFillId() const;
+    int getUniverseFillId() const;
     Universe* getUniverseFill() const;
     int getNumFSRs();
 
-    void setUniverseFill(short int universe_Fill);
+    void setUniverseFill(int universe_Fill);
     void setUniverseFillPointer(Universe* universe_fill);
 
     std::string toString();
